pyproject.toml
[project]
name = "diffsrda"
version = "0.1.0"
description = "Code release for DiffSRDA: diffusion super-resolution data assimilation."
authors = [{ name = "DiffSRDA authors" }]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  "numpy>=1.25,<2.0",
  "scipy>=1.11,<2.0",
  "scikit-learn>=1.3,<2.0",
  "pandas>=2.0,<3.0",
  "tqdm>=4.65,<5.0",
  "pyyaml>=6.0,<7.0",
  "pytz>=2024.1",
  "torch>=2.0,<3.0",
]

[tool.poetry]
packages = [{ include = "src", from = "python" }]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
